% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DZ_combination_gen.R
\name{DZ_combination_gen}
\alias{DZ_combination_gen}
\title{All DZ combinations of all nodes of the hierarchical tree.}
\usage{
DZ_combination_gen(tree.input)
}
\arguments{
\item{tree.input}{the list of two tree structures, one is the simplest
structure with only one layer and the other is the estimated whole tree
structure with multiple layers.}
}
\value{
A list containing all DZ combinations based on estimated tree, as well
as the tree structure with all nodes numbered sequentially.
}
\description{
It creates all DZ combinations based on estimated tree structure.
}
