% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate the classification results}
\usage{
evaluate(vali_set_lable, predict_lable, Indices = NULL)
}
\arguments{
\item{vali_set_lable}{A vector object with the epithelial related genes.}

\item{gene_list}{A vector object with the true cell type lable.}
}
\value{
A list object with Conf = Confusion_matrix, MedF1 = Median_F1_score, F1 = F1_score, Acc = Accuracy, PercUnl = Unlabled_cell_percentage, PopSize = Population_size.
}
\description{
Evaluate the classification results for \code{scDetect} based on predicted cell type and true cell type.
}
\examples{

# Evaluate the classification results


evaluate_results <- evaluate(vali_set_lable,predict_lable)

}
\author{
Yifei Shen
}
