% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_list}
\alias{gene_list}
\title{Gene list used for calculate epithelial score}
\format{
This is a data frame
}
\usage{
gene_list
}
\description{
A list object with the epithelial related genes used for epithelial score calculation.
}
\keyword{datasets}
