% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_position_file}
\alias{gene_position_file}
\title{Gene position used for CNV analysis in tumor cells}
\format{
This is a list
}
\usage{
gene_position_file
}
\description{
A data frame object of gene position with four colums (Chromosome, Start_postition, End_position, Gene_symbol).
}
\keyword{datasets}
