% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{get_cnv_score}
\alias{get_cnv_score}
\title{Generate CNV classification results based on CNV consensus clustering}
\usage{
get_cnv_score(data_matrix, gene_position_file, anno_file_input, output_dir)
}
\arguments{
\item{data_matrix}{A matrix object with genes as rows and cells as columns.}

\item{anno_file_input}{A data frame object of the cell type prediction results based on scDetect.}

\item{output_dir}{A character object with the output directory.}

\item{gene_list}{A list object with the epithelial related genes used for epithelial score calculation.}
}
\value{
A data frame object with "CNV_Class","CNV_entropy_score","anno_file".
}
\description{
Get epithelial score of each cell based on \code{AUCell}.
}
\examples{

# Generate CNV classification results based on CNV consensus clustering


cnv_results <- get_cnv_score(data_matrix,gene_position_file,anno_file_input,output_dir)


}
\author{
Yifei Shen
}
