% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{get_epithelial_score}
\alias{get_epithelial_score}
\title{Get epithelial score of each cell}
\usage{
get_epithelial_score(data_matrix, gene_list)
}
\arguments{
\item{data_matrix}{A matrix object with genes as rows and cells as columns.}

\item{gene_list}{A vector object with the epithelial related genes.}
}
\value{
A vector object with the epithelial of each cell.
}
\description{
Get epithelial score of each cell based on \code{AUCell}.
}
\examples{

# Get epithelial score of each cell


epithelial_score <- get_epithelial_score(data_matrix,gene_list)

}
\author{
Yifei Shen
}
