% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{getCutoff}
\alias{getCutoff}
\title{Get cutoff value for scDetect}
\usage{
getCutoff(train_set_matrix, train_set_lable, feature_list = NULL)
}
\arguments{
\item{train_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_lable}{A vector  object with cell type lables of the cells in train_set_matrix.}

\item{feature_list}{A vector  object with the feature genes generated by \code{getFeature}.}
}
\value{
A list object with cutoff for Intermediate cell type, and cutoff for Unknown cell type.
}
\description{
Get Intermediate-type cutoff value and Unknown cell type cutoff value based on permutation for the scDetect score.
}
\examples{

# Get cutoff value for scDetect

cutoff_results <- getCutoff(train_set_matrix,train_set_lable,feature_list)

}
\author{
Yifei Shen
}
