% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{getFeature}
\alias{getFeature}
\title{Get feature genes from a training dataset}
\usage{
getFeature(train_set_matrix, train_set_lable)
}
\arguments{
\item{train_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_lable}{A vector object with cell type lables of the cells in train_set_matrix.}
}
\value{
A vector object with the feature genes
}
\description{
Get feature genes from a training dataset for \code{scDetect} cell type classification.
}
\examples{

# Get feature genes from a training dataset

feaure_genes <- getFeature(train_set_matrix = train_set_matrix, train_set_lable = train_set_lable)

}
\author{
Yifei Shen
}
