% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.match.R
\name{index.match}
\alias{index.match}
\title{Index of DZ combinations}
\usage{
index.match(dz.combine, all.z.states)
}
\arguments{
\item{dz.combine}{DZ combinations based on estimated tree structure.}

\item{all.z.states}{all DZ combinations based on independent cell types.}
}
\value{
A vector of index.
}
\description{
It matches the index of the DZ combinations based on estimated
tree structures among the all DZ combinations based on independent cell types.
}
