% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDesign_tree.R
\name{makeDesign_tree}
\alias{makeDesign_tree}
\title{Making design matrix}
\usage{
makeDesign_tree(design.1, design.2, factor.to.test)
}
\arguments{
\item{design.1}{covariates representing interested factors to be tested,
such as disease, gender and ethnicity.}

\item{design.2}{other covariates that may also affect expression level
beside the factor of interest.}

\item{factor.to.test}{A phenotype name, e.g. "disease", or a vector of
contrast terms, e.g. c("disease", "case", "control").}
}
\value{
A list containing a design matrix, all covariates of the data
model and their names, the formula of the linear model.
}
\description{
It creates design matrix of covariates of the data model
for gene expression.
}
\examples{
design.1 = data.frame(disease = c(0,1,1,0,0,1,0,0,1,0))
design.2 = data.frame(gender = c(1,0,1,0,0,1,1,0,1,1))
makeDesign_tree(design.1 = design.1, design.2 = design.2,
                             factor.to.test = 'disease')

}
