% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal.L.parallel.R
\name{Marginal.L.parallel}
\alias{Marginal.L.parallel}
\title{Calculation of data likelihood}
\usage{
Marginal.L.parallel(X, Y_raw, geneNames, all.z.states, numCores)
}
\arguments{
\item{X}{the projection matrix of data linear model.}

\item{Y_raw}{A list of matrix of pseudo-bulk data for each cell type.}

\item{geneNames}{names of all genes.}

\item{all.z.states}{all DZ combinations based on independent cell types.}

\item{numCores}{number of cores for parallel running.}
}
\value{
data likelihood
}
\description{
It calculates observed data likelihood based on the data model and
parameter estimation in OLS.
}
