% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma_test}
\alias{melanoma_test}
\title{A SingleCellExperiment with Melanoma test data.}
\format{
This is a SingleCellExperiment
}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}{GEO}
}
\usage{
melanoma_test
}
\description{
A subset of the Melanoma data, from Tirosh et al. (2016), Science.
}
\references{
Tirosh et al. (2016) Science 6282:189-196
}
\keyword{datasets}
