% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muraro_test}
\alias{muraro_test}
\title{Muraro pancreas data: A SingleCellExperiment with celltypes in the "celltypes" colData.}
\format{
This is a SingleCellExperiment
}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE85241}{GEO}
}
\usage{
muraro_test
}
\description{
A subset of the Melanoma data, from Muraro et al. (2016),  Cell Syst.
}
\references{
Muraro et al. (2016),  Cell Syst.
}
\keyword{datasets}
