% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorr.R
\name{plotCorr}
\alias{plotCorr}
\title{Showing DE/DM state correlation between cell types}
\usage{
plotCorr(
  tstat,
  de.state = NULL,
  tstat.thres = NULL,
  p.size = 0.2,
  p.color = grDevices::adjustcolor("black", alpha.f = 0.2),
  fig.margin = c(1, 1, 1, 1),
  fig.margin.unit = "in",
  line.type = "dashed",
  line.color = "blue"
)
}
\arguments{
\item{tstat}{matrix of test statistics from DESeq2.}

\item{de.state}{matrix of DE/DM states (1: DE/DM, 0:non-DE/DM)}

\item{tstat.thres}{threshold of test statistics to define DE/DM,
required if de.state not provided.}

\item{p.size}{point size for scatter plot.}

\item{p.color}{point color for scatter plot.}

\item{fig.margin}{figure margin.}

\item{fig.margin.unit}{unit of figure margin.}

\item{line.type}{line type in scatter plot.}

\item{line.color}{line color in scater plot.}
}
\value{
A figure contains scatter plots, Pearson correlation and
odds ration of test statistics for each pair of cell types.
}
\description{
It generates test statistics for each pair of
cell types and calculate corresponding Pearson correlation
and odds ratio.
}
\examples{
tstat.1 <- runif(1000,0,5)
tstat.2 <- tstat.1 + rnorm(1000,0,1)
tstat.2[tstat.2 < 0] =0
tstat.3 <- runif(1000,0,5)
tstat.input <- data.frame('cell.1'=tstat.1,
                          'cell.2'=tstat.2,
                          'cell.3'=tstat.3)
plotCorr(tstat = tstat.input, tstat.thres = 2.58)
}
