% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_calc_tree.R
\name{post_calc_tree}
\alias{post_calc_tree}
\title{Posterior probability calculation}
\usage{
post_calc_tree(
  Design_matrix,
  Y_raw,
  tree.input,
  DESeq2_res,
  factor.to.test,
  fdr,
  p.value = NULL,
  no_prior_info = FALSE,
  p.matrix.input = NULL,
  core.num = NULL
)
}
\arguments{
\item{Design_matrix}{the design matrix of all the covariates of the data model for
gene expression.}

\item{Y_raw}{A list of matrix of pseudo-bulk data for each cell type.}

\item{tree.input}{the list of two tree structures, one is the simplest
structure with only one layer and the other is the estimated whole
tree structure with multiple layers.}

\item{DESeq2_res}{the results from DESeq2 test in each cell type.}

\item{factor.to.test}{A phenotype name, e.g. "disease", or a vector of contrast terms,
e.g. c("disease", "case", "control").}

\item{fdr}{cut off of fdr used to define DE state to estimate prior prob.}

\item{p.value}{cut off of p-value used to define DE state to estimate prior prob.}

\item{no_prior_info}{A boolean value, whether there is not prior information.}

\item{p.matrix.input}{Prior probability on each node of the tree structure.}

\item{core.num}{number of cores for parallel running.}
}
\value{
A list containing lists for every cell type. Each list contains estimates
of conditional probability for each node, prior probs (weights) for each DZ
combination, all DZ combinations based on estimated tree structure, the hierarchical
tree structure and the posterior probability of each gene in each cell type.
}
\description{
It calculates posterior probability of whether a feature is DE/DM in
each cell type based on Bayesian model.
}
