% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raddlog.matrix.R
\name{Raddlog.matrix}
\alias{Raddlog.matrix}
\title{Getting sum of two probabilities.}
\usage{
Raddlog.matrix(a, b)
}
\arguments{
\item{a}{one single probability.}

\item{b}{another single probability.}
}
\value{
sum of two probabilities.
}
\description{
It describes how to calculate sum of two probabilities.
}
