% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsumlog.matrix.R
\name{Rsumlog.matrix}
\alias{Rsumlog.matrix}
\title{Calculation of marginal probability.}
\usage{
Rsumlog.matrix(a)
}
\arguments{
\item{a}{one single probability.}
}
\value{
the marginal probability.
}
\description{
It calculates marginal probability by summing over all possible conditions.
}
