% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{scDetect}
\alias{scDetect}
\title{Cell type classification using scDetect}
\usage{
scDetect(vali_set_matrix, train_set_matrix, train_set_lable, p_value = NULL)
}
\arguments{
\item{vali_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_lable}{A list object with cell type lables of the cells in train_set_matrix.}

\item{p_value}{A numeric of the cutoff pvalue.}
}
\value{
A data frame object with four colums (predict_lable, predict_score, pvalue, final_predict_lable).
}
\description{
Predict cell types using scDetect based on training data set.
}
\examples{

# Cell type classification using scDetect


prediction_results <- scDetect(vali_set_matrix,train_set_matrix,train_set_lable,feature_list)


}
\author{
Yifei Shen
}
