% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{scDetect_Cancer}
\alias{scDetect_Cancer}
\title{Cell type classification using scDetect-Cancer in tumor scRNA-Seq data}
\usage{
scDetect_Cancer(
  vali_set_matrix,
  train_set_matrix,
  train_set_lable,
  gene_position_file,
  gene_list,
  output_dir
)
}
\arguments{
\item{vali_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_lable}{A list object with cell type lables of the cells in train_set_matrix.}

\item{gene_position_file}{A data frame object of gene position with four colums (Chromosome, Start_postition, End_position, Gene_symbol).}

\item{gene_list}{A list object with the epithelial related genes used for epithelial score calculation.}

\item{output_dir}{A character object with the output directory.}
}
\value{
A list object with a vector "lable" of the prediction cell type, and a data frame "detail_info" of the detailed information.
}
\description{
Predict cell types using scDetect-Canaer based on reference training data set.
}
\examples{

# Cell type classification using scDetect-Cancer in tumor scRNA-Seq data


prediction_cancer_results <- scDetect_Cancer(vali_set_matrix,train_set_matrix,train_set_lable,gene_position_file,gene_list,output_dir)


}
\author{
Yifei Shen
}
