% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDetect.R
\name{scDetect_score}
\alias{scDetect_score}
\title{Calculate scDetect-score}
\usage{
scDetect_score(
  vali_set_matrix,
  train_set_matrix,
  train_set_lable,
  feature_list = NULL
)
}
\arguments{
\item{vali_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_matrix}{A matrix object with genes as rows and cells as columns.}

\item{train_set_lable}{A list object with cell type lables of the cells in train_set_matrix.}

\item{feature_list}{A list object with the feature genes generated by \code{getFeature}.}
}
\value{
A list object with predict_lable and predict_score
}
\description{
Calculate scDetect-score for \code{scDetect} prediction.
}
\examples{

# Calculate scDetect-score


scDetect_score_results <- scDetect_score(vali_set_matrix,train_set_matrix,train_set_lable,feature_list)

}
\author{
Yifei Shen
}
