% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.est.R
\name{tree.est}
\alias{tree.est}
\title{Tree estimation}
\usage{
tree.est(tstat, de.res, similarityFun = NULL, tree.type)
}
\arguments{
\item{tstat}{matrix of test statistics from DESeq2.}

\item{de.res}{prior DE state for tree estimation.}

\item{similarityFun}{Custom function used to calculate similarity between
cell types that used for tree structure estimation.}

\item{tree.type}{Tree type for inference, default is c('single','full').}
}
\value{
A list containing two tree structures, the one is the simplest
tree structure with only one layer and the other is the tree structure
with multiple layers.
}
\description{
It estimates the cell type hierarchical tree structure based
on test statistics from DESeq2 method.
}
