% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W.cal.R
\name{W.cal}
\alias{W.cal}
\title{Calculation of prior probs.}
\usage{
W.cal(
  var.index,
  dz.combine,
  DESeq2_res,
  fdr = 0.05,
  p.value = NULL,
  p.matrix = NULL
)
}
\arguments{
\item{var.index}{the tree structure with all nodes numbered sequentially.}

\item{dz.combine}{all DZ combinations based on estimated tree.}

\item{DESeq2_res}{the results from DESeq2 test in each cell type.}

\item{fdr}{cut off of fdr used to define DE state to estimate prior prob.}

\item{p.value}{cut off of p-value used to define DE state to estimate prior prob.}

\item{p.matrix}{Prior probability on each node of the tree structure.}
}
\value{
A list containing estimates of prior probs (weights) for each DZ combination
and conditional probability for each node.
}
\description{
It estimates prior probs of DE for each node of tree structure, and
calculates joint prior probs for each DZ combination.
}
