% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xin_test}
\alias{xin_test}
\title{Xin pancreas data: A SingleCellExperiment with celltypes in the "celltypes" colData.}
\format{
This is a SingleCellExperiment
}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE81608}{GEO}
}
\usage{
xin_test
}
\description{
A subset of the Melanoma data , from Xin et al. (2016) Cell Metab.
}
\references{
Xin et al. (2016) Cell Metab.
}
\keyword{datasets}
