% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDETECT.R
\name{scDETECT}
\alias{scDETECT}
\title{The main function of scDETECT. It performs scRNA-seq DE analysis
accounting for the cell type correlations.}
\usage{
scDETECT(
  Y_raw,
  design.1,
  design.2 = NULL,
  factor.to.test = NULL,
  pval = NULL,
  p.adj = NULL,
  tree = NULL,
  p.matrix.input = NULL,
  de.state = NULL,
  cutoff.tree = c("fdr", 0.01),
  cutoff.prior.prob = c("pval", 0.01),
  similarity.function = NULL,
  parallel.core = NULL,
  corr.fig = FALSE,
  run.time = TRUE,
  tree.type = c("single", "full")
)
}
\arguments{
\item{Y_raw}{A list of matrix of pseudo-bulk data for each cell type,
with rows representing features and columns representing samples.}

\item{design.1}{covariates representing interested factors to be tested,
such as disease, gender and ethnicity.}

\item{design.2}{other covariates that may also affect expression level
beside the factor of interest.}

\item{factor.to.test}{A phenotype name, e.g. "disease", or a vector of
contrast terms, e.g. c("disease", "case", "control").}

\item{pval}{matrix of p-values from DESeq2.}

\item{p.adj}{matrix of adjusted p-values from DESeq2.}

\item{tree}{hierarchical tree structure used to account cell type correlation.}

\item{p.matrix.input}{Prior probability on each node of the tree structure.}

\item{de.state}{DE/DM state of each feature in each cell type.}

\item{cutoff.tree}{Cut off used to define DE state to estimate tree, it could
be 'fdr', 'pval' or 'tstat'.}

\item{cutoff.prior.prob}{Cut off used to define DE state to estimate prior probs
of nodes on tree, it could be 'fdr' or 'pval'.}

\item{similarity.function}{Custom function used to calculate similarity between
cell types that used for tree structure estimation.}

\item{parallel.core}{Number of cores for parallel running.}

\item{corr.fig}{A boolean value, whether to plot corrrelation between cell types
using function plotCorr().}

\item{run.time}{A boolean value, whether to report running time in seconds.}

\item{tree.type}{Tree type for inference, default is c('single','full').}
}
\value{
A list containing lists for every cell type. Each list contains results
from DESeq2 method, estimates of conditional probability for each node, prior
probs (weights) for each DZ combination, all DZ combinations based on estimated
tree structure, the hierarchical tree structure and the posterior probability
of each gene in each cell type.
}
\description{
It provides posterior probability of whether a feature
is DE in certain cell type given pseudo-bulk data.
}
\examples{
N <- 40 # simulation a dataset with 40 samples
K <- 3 # 3 cell types
C <- 10 # 10 cells are simulated in each sample for each cell type
P <- 500 # 500 features
### simulate single cell data and transform to pseudo-bulk data
design <- data.frame(disease=factor(sample(0:1,size = N,replace=TRUE)))
Y = list()
for (i in 1:K){
  y.pseudo = c()
    for (j in 1:N){
      y = matrix(rnbinom(P*C, size = 1, mu = 1), nrow = P, byrow = FALSE)
      y.pseudo = cbind(y.pseudo, rowSums(y))
    }
  Y[[i]] = y.pseudo
  rownames(Y[[i]]) <- paste0('gene',1:P)
}
res <- scDETECT(Y_raw = Y,
                design.1 = design,
                design.2 = NULL,
                factor.to.test = 'disease',
                cutoff.tree = c('tstat',2.58),
                corr.fig = TRUE,
                cutoff.prior.prob = c('pval',0.1) )
}
